+++ src/iov.c
@@ -176,7 +176,7 @@ void mk_iov_print(struct mk_iov *mk_io)
     char *c;
 
     for (i = 0; i < mk_io->iov_idx; i++) {
-        printf("\n[index=%i len=%i]\n '", i, mk_io->io[i].iov_len);
+        printf("\n[index=%i len=%i]\n '", i, (int) mk_io->io[i].iov_len);
         fflush(stdout);
         for (j=0; j < mk_io->io[i].iov_len; j++) {
             c = mk_io->io[i].iov_base;
